fastapi-backend/pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beanie>=1.30.0",
    "dotenv>=0.9.9",
    "fastapi>=0.115.14",
    "motor>=3.7.1",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "uvicorn>=0.35.0",
    "pdfplumber>=0.11.7",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.20",
    "python-docx>=1.2.0",
    "numpy>=2.3.1",
    "opencv-python>=4.11.0.86",
    "spacy>=3.8.7",
    "pdf2image>=1.17.0",
    "sentence-transformers>=4.1.0",
    "scikit-learn>=1.7.0",
    "groq>=0.29.0",
]
